mixin restructure
