+++ build.gradle.kts
@@ -53,10 +53,10 @@ dependencies {
 	include(implementation(libs.mixinconstraints.get())!!)
     include(implementation(libs.brigadier.kotlin.get())!!)
     include(implementation(libs.ini4j.get())!!)
-    include(implementation("me.saharnooby:qoi-java:1.2.1")!!) // fine. have it your way, because even tho i correctly defined you in the libs.versions.toml, i cant .get() you for some fuckwad reason.
-    include(implementation(libs.qoi.awt.get())!!) // OH BUT THIS ONE I CAN HUUUUUUUH FUCK YOU FUCK YOU CUNT
+    include(implementation(libs.qoi.main.get())!!)
+    include(implementation(libs.qoi.awt.get())!!)
     include(implementation(libs.brotli.get())!!)
-    include(implementation("net.java.dev.jna:jna:5.17.0")!!) // not you tooo.. ugh
+    include(implementation(libs.jna.main.get())!!)
     include(implementation(libs.jna.platform.get())!!)
 
 	modImplementation(libs.grappling.hook)
@@ -67,6 +67,8 @@ dependencies {
 //    modImplementation(libs.yungs.api) // aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
     modImplementation(libs.cloth.config)
 	modImplementation(libs.better.end.island)
+    modImplementation(libs.cicada)
+    modImplementation(libs.`do`.a.barrel.roll)
 
 	modCompileOnly(libs.enderscape)
