+++ src/main/java/one/devos/nautical/winterisms/mixin/client/KeyBindsScreenMixin.java
@@ -1,4 +1,4 @@
-package one.devos.nautical.winterisms.mixin;
+package one.devos.nautical.winterisms.mixin.client;
 
 import com.mojang.blaze3d.platform.InputConstants;
 import net.minecraft.Util;
